requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
crewai>=0.20.0
langchain>=0.0.340
openai>=1.13.3
pdfplumber>=0.10.3
pandas>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=7.0.0